+++ packages/ai/src/batch/batch-types.ts
@@ -4,7 +4,8 @@ import type {
   Experimental_BatchV4Status as BatchV4Status,
   Experimental_BatchLanguageModelV4 as BatchLanguageModelV4,
 } from '@ai-sdk/provider';
-import type { ProviderOptions } from '@ai-sdk/provider-utils';
+import type { ProviderOptions, ToolSet } from '@ai-sdk/provider-utils';
+import type { ContentPart } from '../generate-text/content-part';
 import type { LanguageModelCallOptions } from '../prompt/language-model-call-options';
 import type { Prompt } from '../prompt/prompt';
 import type {
@@ -104,9 +105,17 @@ export type BatchOperationOptions = {
 } & BatchRequestOptions;
 
 /**
+ * A Core-normalized content part returned for a successful text batch item.
+ *
+ * Batch requests do not carry a request-specific typed tool set.
+ */
+export type BatchContentPart = ContentPart<ToolSet>;
+
+/**
  * A normalized result for a successful text batch item.
  */
 export type TextBatchGenerationResult = {
+  readonly content: BatchContentPart[];
   readonly text: string;
   readonly finishReason: FinishReason;
   readonly rawFinishReason?: string;
