+++ packages/ai/src/batch/batch.ts
@@ -4,9 +4,13 @@ import {
   type Experimental_BatchV4ItemResult as BatchV4ItemResult,
   type LanguageModelV4,
   type LanguageModelV4GenerateResult,
+  type LanguageModelV4ToolCall,
 } from '@ai-sdk/provider';
-import { withUserAgentSuffix } from '@ai-sdk/provider-utils';
+import { withUserAgentSuffix, type ToolSet } from '@ai-sdk/provider-utils';
 import { InvalidArgumentError } from '../error/invalid-argument-error';
+import { asContent } from '../generate-text/as-content';
+import { parseToolCall } from '../generate-text/parse-tool-call';
+import type { TypedToolCall } from '../generate-text/tool-call';
 import { logWarnings } from '../logger/log-warnings';
 import { resolveLanguageModel } from '../model/resolve-model';
 import { convertToLanguageModelPrompt } from '../prompt/convert-to-language-model-prompt';
@@ -21,6 +25,7 @@ import { prepareRetries } from '../util/prepare-retries';
 import { VERSION } from '../version';
 import type {
   BatchOperationOptions,
+  BatchContentPart,
   BatchReference,
   BatchStatus,
   StartTextBatchOptions,
@@ -175,8 +180,8 @@ export function getBatchResults({
     BatchV4ItemResult<LanguageModelV4GenerateResult>,
     TextBatchItemResult
   > & { cancel?: (reason?: unknown) => void } = {
-    transform(item, controller) {
-      controller.enqueue(convertBatchItemResult(item));
+    async transform(item, controller) {
+      controller.enqueue(await convertBatchItemResult(item));
     },
 
     cancel(reason) {
@@ -296,9 +301,9 @@ function validateBatchReference({
   }
 }
 
-function convertBatchItemResult(
+async function convertBatchItemResult(
   item: BatchV4ItemResult<LanguageModelV4GenerateResult>,
-): TextBatchItemResult {
+): Promise<TextBatchItemResult> {
   if (item.status !== 'succeeded') {
     return item;
   }
@@ -306,14 +311,15 @@ function convertBatchItemResult(
   return {
     id: item.id,
     status: 'succeeded',
-    ...convertGenerateResult(item.result),
+    ...(await convertGenerateResult(item.result)),
   };
 }
 
-function convertGenerateResult(
+async function convertGenerateResult(
   result: LanguageModelV4GenerateResult,
-): TextBatchGenerationResult {
+): Promise<TextBatchGenerationResult> {
   return {
+    content: await convertContent(result.content),
     text: result.content
       .filter(
         (part): part is Extract<typeof part, { type: 'text' }> =>
@@ -335,4 +341,33 @@ function convertGenerateResult(
       : {}),
     providerMetadata: result.providerMetadata,
   };
+}
+
+async function convertContent(
+  content: LanguageModelV4GenerateResult['content'],
+): Promise<Array<BatchContentPart>> {
+  const toolCalls: Array<TypedToolCall<ToolSet>> = await Promise.all(
+    content
+      .filter(
+        (part): part is LanguageModelV4ToolCall => part.type === 'tool-call',
+      )
+      .map(toolCall =>
+        parseToolCall<ToolSet>({
+          toolCall,
+          tools: undefined,
+          repairToolCall: undefined,
+          messages: [],
+          instructions: undefined,
+        }),
+      ),
+  );
+
+  return asContent<ToolSet>({
+    content,
+    toolCalls,
+    toolOutputs: [],
+    toolApprovalRequests: [],
+    toolApprovalResponses: [],
+    tools: undefined,
+  });
 }
